rename util/ folder to utils/ for clarity and consistency
