+++ lib/data/services/export/web_absence_i_cal_exporter.dart
@@ -1,6 +1,6 @@
 import 'package:absence_manager/domain/models/absence_with_member.dart';
 import 'package:absence_manager/domain/use_cases/absence_i_cal_exporter.dart';
-import 'package:absence_manager/util/i_cal_exporter.dart';
+import 'package:absence_manager/utils/i_cal_exporter.dart';
 import 'package:universal_html/html.dart' as html;
 
 /// Web implementation of [AbsenceExporter].
